Fix tagorder when the argument contains a pattern
If the argument does not start with '/' then always treat it as a word
literal by escaping it properly for taglist. This behaviour is
consistent with ":tag" and ":stag", meaning that the jump will actually go
to the correct tag location. To get the best performance it is important
that the pattern taglist starts with exactly '^', else the binary search
seems to be skipped.

When argument is a pattern starting with '/', don't add the exact match
match anchors '^' and '$'.

:help taglist
taglist({expr}) Returns a list of tags matching the regular expression
{expr}

:help tag-regexp
The ":tag" and ":tselect" commands accept a regular expression argument.
When the argument starts with '/', it is used as a pattern.  If the argument
does not start with '/', it is taken literally, as a full tag name.